Finished the logic in the backend and started writing the results.html file
I wrote the getSubmittions function in the main.go file with debug
printing the data that is supposed to reach the results.html file.
Started writing the results.html file by adding all of the headers, some
CSS and some of the logic needed to deal with the recieved data from the
go server.